Moved project files to a separate folder
