+++ ConsoleHabitLogger.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34511.84
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "habit-logger", "habit-logger.csproj", "{E5BD22C0-254E-4023-81A2-79191B4F17B4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleHabitLogger", "ConsoleHabitLogger\ConsoleHabitLogger.csproj", "{E5BD22C0-254E-4023-81A2-79191B4F17B4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
