+++ ConsoleHabitLogger/Locale/es.yaml
