--- Utils.cs
@@ -1,24 +1,0 @@
-namespace habit_logger 
-{
-	static class Utils
-	{
-		// Safe alternative to use instead of Console.Clear() to fix problems with Visual Studio Code debugger.
-		public static void ConsoleClear()
-		{
-			try 
-			{
-				Console.Clear();
-			} 
-			catch (IOException) 
-			{
-				Console.WriteLine("IOException - you can ignore it if you're using Visual Studio Code debugger.\n");
-			}
-		}
-
-        public static void PressAnyKeyToContinue()
-        {
-            Console.WriteLine("Press any key to continue.");
-            Console.ReadKey();
-        }
-	}
-}
