--- global.json
@@ -1,7 +1,0 @@
-{
-  "sdk": {
-    "version": "8.0.0",
-    "rollForward": "latestMajor",
-    "allowPrerelease": true
-  }
-}
